# docs/openapi.yaml
openapi: 3.0.3
info:
  title: mStream API
  version: "1"
  description: |
    REST API for the [mStream](https://github.com/IrosTheBeggar/mStream) music streaming server.

    ## Authentication

    Most endpoints require a JWT bearer token. Tokens are issued by `POST /api/v1/auth/login` and
    can be supplied via any of (checked in order):

    1. Request body field `token`
    2. Query parameter `token`
    3. HTTP header `x-access-token`
    4. Cookie `x-access-token`

    ### Public mode

    If no users exist in the database, all requests are authenticated as a virtual public user with
    full library access — no token required. This is the default for fresh installs.

    ### Special tokens

    - **Admin tokens** have `admin: true` in the payload and are required for every `/api/v1/admin/*` endpoint.
    - **Jukebox tokens** (`jukebox: true`) are short-lived and bound to an active remote-control session.
    - **Shared-playlist tokens** (`shareToken: true`) restrict file access to the tracks in one specific shared playlist.
    - **Federation invite tokens** (`invite: true`) can only call `POST /api/v1/federation/invite/exchange`.

    ## Errors

    Errors use the shape `{ "error": "<message>" }`. Status codes:

    | Code | Meaning |
    |------|---------|
    | 400  | Bad request (invalid input) |
    | 401  | Unauthenticated |
    | 403  | Forbidden (Joi validation error, missing permission) |
    | 404  | Resource not found |
    | 405  | Admin API locked |
...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass the JWT in the `x-access-token` header. Despite the OpenAPI "bearer" label,
        the value does NOT include a `Bearer ` prefix — it's the raw token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: x-access-token
...
paths:
...
  /api/v1/admin/db/params:
    get:
      tags: [Admin - Scanner]
      summary: Get scanner parameters
      responses:
        "200":
          description: Current scanner settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanOptions" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/scan-interval:
    post:
      tags: [Admin - Scanner]
      summary: Set scan interval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scanInterval]
              properties:
                scanInterval: { type: integer, minimum: 0, description: "Hours. 0 disables scheduled rescans." }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/skip-img:
    post:
      tags: [Admin - Scanner]
      summary: Set whether album art extraction is skipped
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skipImg]
              properties:
                skipImg: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/boot-scan-delay:
    post:
      tags: [Admin - Scanner]
      summary: Set boot-scan delay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bootScanDelay]
              properties:
                bootScanDelay: { type: integer, minimum: 0, description: "Seconds." }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/max-concurrent-scans:
    post:
      tags: [Admin - Scanner]
      summary: Set max concurrent scans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [maxConcurrentTasks]
              properties:
                maxConcurrentTasks: { type: integer, minimum: 0 }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/compress-image:
    post:
      tags: [Admin - Scanner]
      summary: Set whether album art thumbnails are generated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [compressImage]
              properties:
                compressImage: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/scan-commit-interval:
    post:
      tags: [Admin - Scanner]
      summary: Set scanner commit cadence
      description: Number of files between DB commits during a scan. Lower = shorter write-lock holds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scanCommitInterval]
              properties:
                scanCommitInterval: { type: integer, minimum: 1 }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/generate-waveforms:
    post:
      tags: [Admin - Scanner]
      summary: Toggle post-scan waveform generation
      description: |
        When `false`, the bulk waveform generator is skipped after scans finish.
        Waveforms can still be produced on-demand by `GET /api/v1/db/waveform`
        (which reads the cache and only calls ffmpeg on a miss). Useful on
        CPU-constrained Docker hosts where the periodic generator spike is
        visible in `top`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [generateWaveforms]
              properties:
                generateWaveforms: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/auto-album-art:
    post:
      tags: [Admin - Scanner]
      summary: Toggle auto album art lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [autoAlbumArt]
              properties:
                autoAlbumArt: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/album-art-write-to-folder:
    post:
      tags: [Admin - Scanner]
      summary: Write fetched cover art as `cover.jpg` in the album folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [albumArtWriteToFolder]
              properties:
                albumArtWriteToFolder: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/album-art-write-to-file:
    post:
      tags: [Admin - Scanner]
      summary: Embed fetched cover art in audio files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [albumArtWriteToFile]
              properties:
                albumArtWriteToFile: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/params/album-art-services:
    post:
      tags: [Admin - Scanner]
      summary: Select which services to query for album art
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [albumArtServices]
              properties:
                albumArtServices:
                  type: array
                  items:
                    type: string
                    enum: [musicbrainz, itunes, deezer]
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/force-compress-images:
    post:
      tags: [Admin - Scanner]
      summary: Re-run the album-art thumbnail compressor over the whole library
      responses:
        "200":
          description: Job status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  started: { type: boolean }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/generate-waveforms:
    post:
      tags: [Admin - Scanner]
      summary: Generate waveform data for every track
      responses:
        "200":
          description: Job status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  started: { type: boolean }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/scan/all:
    post:
      tags: [Admin - Scanner]
      summary: Scan all libraries
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/scan/force-rescan:
    post:
      tags: [Admin - Scanner]
      summary: Rescan all libraries ignoring mtimes
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/db/scan/stats:
    get:
      tags: [Admin - Scanner]
      summary: Total track count in the DB
      responses:
        "200":
          description: Count.
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileCount: { type: integer }
        "405": { $ref: "#/components/responses/AdminLocked" }
...
  /api/v1/admin/users:
    get:
      tags: [Admin - Users]
      summary: List all users
      responses:
        "200":
          description: Users keyed by username.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { $ref: "#/components/schemas/User" }
        "405": { $ref: "#/components/responses/AdminLocked" }
    put:
      tags: [Admin - Users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, vpaths]
              properties:
                username: { type: string }
                password: { type: string }
                vpaths:
                  type: array
                  items: { type: string }
                admin: { type: boolean, default: false }
                allowMkdir: { type: boolean, default: true }
                allowUpload: { type: boolean, default: true }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
    delete:
      tags: [Admin - Users]
      summary: Delete a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/users/password:
    post:
      tags: [Admin - Users]
      summary: Change a user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/users/lastfm:
    post:
      tags: [Admin - Users]
      summary: Set Last.fm credentials for a user
      description: Note — the field names contain a typo (`lasftfm*`) that is preserved for backwards compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, lasftfmUser, lasftfmPassword]
              properties:
                username: { type: string }
                lasftfmUser: { type: string }
                lasftfmPassword: { type: string }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/users/vpaths:
    post:
      tags: [Admin - Users]
      summary: Update a user's library access list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, vpaths]
              properties:
                username: { type: string }
                vpaths:
                  type: array
                  items: { type: string }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
  /api/v1/admin/users/access:
    post:
      tags: [Admin - Users]
      summary: Update a user's role / permission flags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, admin, allowMkdir, allowUpload]
              properties:
                username: { type: string }
                admin: { type: boolean }
                allowMkdir: { type: boolean }
                allowUpload: { type: boolean }
                allowFileModify: { type: boolean, default: true }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }

  # ── Admin - Libraries ────────────────────────────────────────────────────
  /api/v1/admin/directories:
    get:
      tags: [Admin - Libraries]
      summary: List all libraries
      description: |
        Returns one of two shapes depending on caller:
        - Legacy (admin UI): `{ [vpath]: { root, type } }`
        - Velvet UI: array of `{ name, root, type }` when fetched via the Velvet route.
      responses:
        "200":
          description: Libraries.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    additionalProperties:
                      type: object
                      properties:
                        root: { type: string }
                        type: { type: string }
                  - type: array
                    items: { $ref: "#/components/schemas/Library" }
        "405": { $ref: "#/components/responses/AdminLocked" }
...
  /api/v1/admin/lock-api:
    post:
      tags: [Admin - Config]
      summary: Lock or unlock the admin API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lock]
              properties:
                lock: { type: boolean }
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }
...
  /api/v1/admin/transcode/download:
    post:
      tags: [Admin - Transcode]
      summary: Download the bundled ffmpeg binary
      responses:
        "200": { $ref: "#/components/responses/EmptySuccess" }
        "405": { $ref: "#/components/responses/AdminLocked" }

  # ── Admin - DLNA ─────────────────────────────────────────────────────────
...